refactor(settings): extract shared row primitives into components/SettingsPrimitives.kt
Move Header, SettingsSection, SettingsSubheader, SettingRow, ChoiceRow
(+ ChoiceDialog), SwitchRow, SliderRow, ActionRow, ResetRow (+
ResetConfirmDialog), TrailingIcon, FontRow, and the
Modifier.clipClickable helper out of SettingsScreen.kt into a new
components/SettingsPrimitives.kt, under a new
io.github.seijikohara.femto.ui.settings.components package.

This is a pure structural move ahead of splitting SettingsScreen.kt
into one file per section: those upcoming section files live in the
same new package and call these shared row primitives, which needs
internal (module-wide) visibility -- private is file-scoped and
cannot reach across the package split. Widen the 11 declarations
every section file will call to internal; ChoiceDialog,
ResetConfirmDialog, and clipClickable stay private since each is
only ever called by the one sibling declaration it moved in with.

No behavior change: every call site in SettingsScreen.kt is
unchanged, only the declaration location and 11 visibility modifiers
moved. Also drops the SettingsScreen.kt imports that only the moved
declarations needed.
